Extract forward declarations
